added authentication to all links

+++ lab3/lab3b/index.js
@@ -169,7 +169,6 @@ app.get('/courtroom', function(req, res) {
 
 // post new courtroom
 app.post('/courtroom', function(req, res) {
-
   var authentication = authsplit(req.get('Authentication') || '');
 
   if(!isAuthentication(authentication)) {
@@ -200,6 +199,15 @@ app.get('/courtroom/:id', function(req, res) {
 
 // delete a single courtroom
 app.delete('/courtroom/:id', function(req, res) {
+  var authentication = authsplit(req.get('Authentication') || '');
+
+  if(!isAuthentication(authentication)) {
+    return res.status('401').send( {
+      valid: false,
+      message: "invalid token!"
+    })
+  }
+
   models.Courtroom.destroy({
     where: {
       id: req.params.id
@@ -212,6 +220,15 @@ app.delete('/courtroom/:id', function(req, res) {
 
 // update single courtroom
 app.put('/courtroom/:id', function(req, res) {
+  var authentication = authsplit(req.get('Authentication') || '');
+
+  if(!isAuthentication(authentication)) {
+    return res.status('401').send( {
+      valid: false,
+      message: "invalid token!"
+    })
+  }
+
   models.Courtroom.find({
     where: {
       id: req.params.id
@@ -240,6 +257,15 @@ app.get('/judge', function(req, res) {
 
 // post new judge
 app.post('/judge', function(req, res) {
+  var authentication = authsplit(req.get('Authentication') || '');
+
+  if(!isAuthentication(authentication)) {
+    return res.status('401').send( {
+      valid: false,
+      message: "invalid token!"
+    })
+  }
+
   models.Judge.create({
     name: req.body.name,
     room: req.body.room,
@@ -264,6 +290,15 @@ app.get('/judge/:id', function(req, res) {
 
 // delete a single judge
 app.delete('/judge/:id', function(req, res) {
+  var authentication = authsplit(req.get('Authentication') || '');
+
+  if(!isAuthentication(authentication)) {
+    return res.status('401').send( {
+      valid: false,
+      message: "invalid token!"
+    })
+  }
+
   models.Judge.destroy({
     where: {
       id: req.params.id
@@ -276,6 +311,15 @@ app.delete('/judge/:id', function(req, res) {
 
 // update single judge
 app.put('/judge/:id', function(req, res) {
+  var authentication = authsplit(req.get('Authentication') || '');
+
+  if(!isAuthentication(authentication)) {
+    return res.status('401').send( {
+      valid: false,
+      message: "invalid token!"
+    })
+  }
+
   models.Judge.find({
     where: {
       id: req.params.id
@@ -306,6 +350,15 @@ app.get('/participant', function(req, res) {
 
 // post new participant
 app.post('/participant', function(req, res) {
+  var authentication = authsplit(req.get('Authentication') || '');
+
+  if(!isAuthentication(authentication)) {
+    return res.status('401').send( {
+      valid: false,
+      message: "invalid token!"
+    })
+  }
+
   models.Participant.create({
     name: req.body.name,
     address: req.body.address,
@@ -329,6 +382,15 @@ app.get('/participant/:id', function(req, res) {
 
 // delete a single participant
 app.delete('/participant/:id', function(req, res) {
+  var authentication = authsplit(req.get('Authentication') || '');
+
+  if(!isAuthentication(authentication)) {
+    return res.status('401').send( {
+      valid: false,
+      message: "invalid token!"
+    })
+  }
+
   models.Participant.destroy({
     where: {
       id: req.params.id
@@ -341,6 +403,15 @@ app.delete('/participant/:id', function(req, res) {
 
 // update single participant
 app.put('/participant/:id', function(req, res) {
+  var authentication = authsplit(req.get('Authentication') || '');
+
+  if(!isAuthentication(authentication)) {
+    return res.status('401').send( {
+      valid: false,
+      message: "invalid token!"
+    })
+  }
+
   models.Participant.find({
     where: {
       id: req.params.id
@@ -371,6 +442,14 @@ app.get('/case', function(req, res) {
 
 // post new case
 app.post('/case', function(req, res) {
+  var authentication = authsplit(req.get('Authentication') || '');
+
+  if(!isAuthentication(authentication)) {
+    return res.status('401').send( {
+      valid: false,
+      message: "invalid token!"
+    })
+  }
   models.Case.create({
     judge_id: req.body.judge_id,
     courtroom_id: req.body.courtroom_id,
@@ -398,6 +477,14 @@ app.get('/case/:id', function(req, res) {
 
 // delete a single case
 app.delete('/case/:id', function(req, res) {
+  var authentication = authsplit(req.get('Authentication') || '');
+
+  if(!isAuthentication(authentication)) {
+    return res.status('401').send( {
+      valid: false,
+      message: "invalid token!"
+    })
+  }
   models.Case.destroy({
     where: {
       id: req.params.id
@@ -410,6 +497,14 @@ app.delete('/case/:id', function(req, res) {
 
 // update single case
 app.put('/case/:id', function(req, res) {
+  var authentication = authsplit(req.get('Authentication') || '');
+
+  if(!isAuthentication(authentication)) {
+    return res.status('401').send( {
+      valid: false,
+      message: "invalid token!"
+    })
+  }
   models.Case.find({
     where: {
       id: req.params.id
